# Faster CKKS multiply