Speeds up CKKS Multiply by ~1.2x - 2.1x. Several optimizations improve the default implementation:
- Avoid the `add_poly_coeffmod` step when not needed
- Use as much in-place computation as possible
- Process 256-coefficient tiles at a time rather than entire polynomials at a time.
- Also runs pre-commit for some whitespace changes

I imagine a similar approach could be used to optimize the [BEHZ ciphertext product](https://github.com/microsoft/SEAL/blob/main/native/src/seal/evaluator.cpp#L418-L436) in BFV multiply.

If this approach turns out to improve performance on more evaluator functions, perhaps a better implementation would be to add another iterator layer beyond the CoeffIterator, e.g. "TileIterator", which performs the tiling iteration more cleanly.

On ICX compiled with gcc-10, running `CKKS / EvaluateMulCt` benchmark for 1000 iterations yields:

n=4096

| HEXL | Time before (us) | Time after (us) | Speedup |
|------|------------------|-----------------|---------|
| OFF  |  148             | 120             | 1.23x   |
| ON   |  52.6            | 39.1            | 1.34x   |

n=8192

| HEXL | Time before (us) | Time after (us) | Speedup |
|------|------------------|-----------------| ------- |
| OFF  |  640             | 493             | 1.30x.  |
| ON   |  209             | 116             | 1.82x.  |

n=16384

| HEXL | Time before (us) | Time after (us) | Speedup |
|------|------------------|-----------------| ------- |
| OFF  | 2656            | 2008             | 1.32x   |
| ON   |  891             | 414             | 2.15x   |

n=32768

| HEXL | Time before (us) | Time after (us) | Speedup |
|------|------------------|-----------------| ------- |
| OFF  | 10054            | 7634            | 1.31x   |
| ON   |  3963            | 2509            | 1.57x   |