@fionser 2021-05-28
- CKKS multiplication ? BFV multiplication ?
- Clang Formatting seems not working?