@fionser 2021-05-28
Typos in the comments
```
 // Given input pairs of polynomials x = (x[0], x[1], x[2]), y = (y[0], y[1])
 ```
Should it be `x = (x[0], x[1])` ?