@fboemer 2021-05-28
> Typos in the comments
> 
> ```
>  // Given input pairs of polynomials x = (x[0], x[1], x[2]), y = (y[0], y[1])
> ```
> 
> Should it be `x = (x[0], x[1])` ?

The input to CKKS multiply would be `x = (x[0], x[1])`, but after `encrypted1.resize(context_, context_data.parms_id(), dest_size);`, `x = (x[0], x[1], x[2])`. So I suppose `x = (x[0], x[1], x[2])` is a tuple, rather than a pair, of polynomials. Hopefully that makes it more clear.