@WeiDaiWD 2021-06-08
This is quite nice! Thanks a lot. I'm gonna make slight modifications before release (3.6.6).